backend/explainer/requirements.txt
python-pptx==0.6.21
google-generativeai==0.3.1
python-dotenv==1.0.0
asyncio==3.4.3
sqlalchemy==2.0.23